[halo] [wip] major refactor of perf_events code
this refactor was needed because we need to set
the perf_events inherit bit so that threads
spawned in the process after intialization
are still tracked by perf events.
In order to set this bit and track threads across all CPUs,
we need to maintain a perf_event handle for each CPU,
 instead of just one handle that the code was initially
designed for.